Reorganizes files into new component folder
