+++ get-in-line/src/App.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import LoginPage from './LoginPage';
+import LoginPage from './components/LoginPage';
 
 function App() {
   return (
